requirements.txt
openai>=1.30.0
chromadb>=0.5.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
langchain>=0.2.0
langchain-community>=0.2.0
pypdf>=4.0.0
python-dotenv>=1.0.0
streamlit>=1.35.0
datasets>=2.19.0
numpy>=1.26.0